requirements.txt
pyTelegramBotApi
Pillow
flask
gunicorn
 
